(refactor) - Refactors scoring formula into seperate module
